Adjusted movement speed using application time to adjust for VSYNC

+++ DirectX 11 App/FrameTimer.h
@@ -12,7 +12,7 @@ public:
 	void Reset();
 	bool IsActive() const { return active; }
 
-	double GetMilliseconds();
+	float GetMilliseconds();
 
 	static void IncrementFPSCounter();
 	static void ResetFPSCounter();
